// apps/web/test/__generated__/RelayPaginationSubjectRefetchQuery.graphql.ts
export type RelayPaginationSubjectRefetchQuery$variables = {
  count: number;
  cursor?: any | null | undefined;
  id: string;
};
export type RelayPaginationSubjectRefetchQuery$data = {
  readonly node: {
    readonly " $fragmentSpreads": FragmentRefs<"RelayPaginationSubjectFragment">;
  } | null | undefined;
};
export type RelayPaginationSubjectRefetchQuery = {
  response: RelayPaginationSubjectRefetchQuery$data;
  variables: RelayPaginationSubjectRefetchQuery$variables;
};